Changed the placement for advanced button
