+++ src/pages/FriendsAndFamily/EnterGiftDetails.tsx
@@ -9,7 +9,8 @@ import {
   StatusBar,
   ScrollView,
   Platform,
-  ImagePropTypes
+  FlatList,
+  TouchableWithoutFeedback
 } from 'react-native'
 import {
   widthPercentageToDP as wp,
@@ -22,7 +23,7 @@ import { RFValue } from 'react-native-responsive-fontsize'
 import FontAwesome from 'react-native-vector-icons/FontAwesome'
 import HeaderTitle from '../../components/HeaderTitle'
 import CommonStyles from '../../common/Styles/Styles'
-import { Gift, GiftThemeId, Wallet } from '../../bitcoin/utilities/Interface'
+import { Gift, GiftThemeId, Wallet, DeepLinkEncryptionType } from '../../bitcoin/utilities/Interface'
 import idx from 'idx'
 import CheckMark from '../../assets/images/svgs/checkmark.svg'
 import AccountShell from '../../common/data/models/AccountShell'
@@ -39,9 +40,67 @@ import { updateUserName } from '../../store/actions/storage'
 
 import { translations } from '../../common/content/LocContext'
 
+import RadioButton from '../../components/RadioButton'
+import Feather from 'react-native-vector-icons/Feather'
+import ModalContainer from '../../components/home/ModalContainer'
+import { AppBottomSheetTouchableWrapper } from '../../components/AppBottomSheetTouchableWrapper'
+
+const ADVANCEDSETTINGDATA = [
+  {
+    id: '1',
+    title: 'F&F Identification',
+    subtitle: 'Use a phone number or email ID stored in your contact details. The recipient needs to confirm their own details',
+  },
+  {
+    id: '2',
+    title: 'Simple OTP',
+    subtitle: 'A 6 digit OTP to be sent separately to the recipient. Good for reducing the chances of man-in-the-middle attack',
+  },
+  {
+    id: '3',
+    title: 'Long OTP',
+    subtitle: 'Unguessable long OTP to be sent tot he recipient separately. Improved security against server hacks and man-in-the-middle',
+  },
+  {
+    id: '4',
+    title: 'Custom Secret Phrase',
+    subtitle: 'Manually provide a passphrase with a hint that the recipient has to confirm',
+  },
+  {
+    id: '5',
+    title: 'No Second Factor',
+    subtitle: 'Good for small gifts',
+  },
+];
+
+
+const FNFIDENTIFICATIONDATA = [
+  {
+    id: '1',
+    title: 'Confirm with phone number',
+    subtitle: 'The recipient will have to confirm their phone number <phone number> to accept the gift',
+  },
+  {
+    id: '2',
+    title: 'Confirm with email address',
+    subtitle: 'The recipient will have to confirm their email address <email address> to accept the gift',
+  },
+];
+
 const GiftDetails = ( { navigation } ) => {
+
+
+  const renderItem = ( { item } ) => (
+    <SettingCard id={item.id} title={item.title} subTitle={item.subtitle} />
+  )
+
+
+  const renderIdentificationItem = ({ item }) => (
+    <IdentificationCard id={item.id} title={item.title} subtitle={item.subtitle} />
+  );
+
   const dispatch = useDispatch()
-  const { giftId, contact, encryptionType } = navigation.state.params
+  const { giftId, contact } = navigation.state.params
   const wallet: Wallet = useSelector( state => state.storage.wallet )
   const strings = translations[ 'f&f' ]
   // const login = translations[ 'login' ]
@@ -51,7 +110,17 @@ const GiftDetails = ( { navigation } ) => {
   const [ dropdownBoxOpenClose, setDropdownBoxOpenClose ] = useState( false )
   const [ addfNf, setAddfNf ] = useState( false )
   const [ dropdownBoxList, setDropdownBoxList ] = useState( [] )
-  const [ isDisabled, setIsDisabled ] = useState( false )
+  const [ advanceSettingsModal, setAdvanceSettingsModal ] = useState( false )
+  const [ selectedAdvancedOptionId, setSelectedAdvancedOptionId ] = useState( '1' )
+  const [ FnFIdentificationModal, setFnFIdentificationModal ] = useState( false )
+  const [ selectedFAndFId, setSelectedFAndFId ] = useState( '1' )
+
+  const [ customSecretIdentificationModal, setCustomSecretIdentificationModal ] = useState( false )
+
+  const [secretPhrase, setSecretPhrase] = useState(true)
+  const [confirmSecretPhrase, setconfirmSecretPhrase] = useState(true)
+  const [encryptionType, setEncryptionType] = useState( DeepLinkEncryptionType.DEFAULT)
+
   const [ dropdownBoxValue, setDropdownBoxValue ] = useState( {
     id: GiftThemeId.ONE,
     title: 'Gift Sats',
@@ -66,6 +135,345 @@ const GiftDetails = ( { navigation } ) => {
   }, [] )
 
   const { title, walletName, gift, avatar }: {title: string, walletName: string, gift: Gift, avatar: boolean} = navigation.state.params
+
+
+  const IdentificationCard = ( { id, title, subtitle } ) => {
+    return (
+      <AppBottomSheetTouchableWrapper
+        onPress={() => {setSelectedFAndFId( id )}}>
+
+        <View style={styles.cardContainer}>
+          <View style={styles.radioBtnContainer}>
+            <RadioButton
+              isChecked={id === selectedFAndFId}
+              size={20}
+              color={Colors.lightBlue}
+              borderColor={Colors.borderColor}
+            />
+          </View>
+          <View>
+            <Text style={styles.identificationHeading}>{title}</Text>
+            <Text numberOfLines={2} style={styles.identificationDescription}>{subtitle}</Text>
+          </View>
+
+        </View>
+      </AppBottomSheetTouchableWrapper>
+    )
+  }
+
+  const selectAdvancedOption = ( id: string ) => {
+    setSelectedAdvancedOptionId( id )
+    setAdvanceSettingsModal( false )
+
+    switch( id ){
+        case '1':
+          setEncryptionType( DeepLinkEncryptionType.DEFAULT )
+          break
+
+        case '2':
+          setFnFIdentificationModal( true ) // selected F&F
+          break
+
+        case '3':
+          setEncryptionType( DeepLinkEncryptionType.OTP )
+          break
+
+        case '4':
+          setEncryptionType( DeepLinkEncryptionType.LONG_OTP )
+          break
+
+        case '5':
+          setCustomSecretIdentificationModal( true )
+          break
+
+        default:
+          setEncryptionType( DeepLinkEncryptionType.DEFAULT )
+    }
+  }
+
+  const SettingCard = ( { id, title, subTitle } ) => {
+    return (
+      <AppBottomSheetTouchableWrapper
+        onPress={() => {
+          selectAdvancedOption( id )
+        }}>
+        <View style={styles.cardContainer}>
+          <View style={styles.radioBtnContainer}>
+            <RadioButton
+              isChecked={id === selectedAdvancedOptionId }
+              size={20}
+              color={Colors.lightBlue}
+              borderColor={Colors.borderColor}
+              onpress={() => {selectAdvancedOption( id )}}
+            />
+          </View>
+          <View>
+            <Text style={styles.identificationHeading}>{title}</Text>
+            <Text numberOfLines={2} style={styles.identificationDescription}>{subTitle}</Text>
+          </View>
+        </View>
+      </AppBottomSheetTouchableWrapper>
+    )
+  }
+
+    const AdvancedSettingsModal = () => {
+      return (
+        <View style={styles.modalContentContainer}>
+          <TouchableOpacity
+            activeOpacity={1}
+            onPress={() => {
+              setAdvanceSettingsModal(false);
+            }}
+            style={{
+              width: wp(7),
+              height: wp(7),
+              borderRadius: wp(7 / 2),
+              alignSelf: 'flex-end',
+              backgroundColor: Colors.lightBlue,
+              alignItems: 'center',
+              justifyContent: 'center',
+              marginTop: wp(3),
+              marginRight: wp(3),
+            }}
+          >
+            <FontAwesome name="close" color={Colors.white} size={19} />
+          </TouchableOpacity>
+          <View>
+            <View
+              style={{
+                marginLeft: wp(7),
+              }}
+            >
+              <Text style={{ ...styles.modalTitleText, fontSize: 18, fontFamily: Fonts.FiraSansRegular }}>Add Second Factor</Text>
+              <Text
+                style={{
+                  ...styles.modalInfoText,
+                  fontFamily: Fonts.FiraSansRegular,
+                  fontSize: 14,
+                }}
+              >
+                {'Add a second factor to your Gift Sats that is generated on the App. For identity Confirmation or Improving security'}
+              </Text>
+            </View>
+
+            <Text
+              style={{
+                ...styles.modalInfoText,
+                paddingTop: 8,
+                marginLeft: 30,
+                fontFamily: Fonts.FiraSansRegular,
+                fontSize: 14,
+              }}
+            >
+              {'The gift can be accessed without providing any second factor. Good for small gifts'}
+            </Text>
+
+            <FlatList data={ADVANCEDSETTINGDATA} renderItem={renderItem} keyExtractor={(item) => item.id} />
+            <Text style={{
+              ...styles.modalInfoText,
+              paddingTop: 8,
+              marginLeft: 30,
+              fontFamily: Fonts.FiraSansRegular,
+              fontSize: 14,
+              marginTop: 50
+            }}>The option selected above will be used to encrypt your Gift Sats. Don't use the medium/ app for communicating the 2nd factor that you use to send the gift link/QR</Text>
+          </View>
+        </View>
+      );
+    }
+
+
+    const FandFIndentificationModal = () => {
+      return (
+        <View style={styles.modalContentContainer}>
+          <TouchableOpacity
+            activeOpacity={1}
+            onPress={() => {
+              setFnFIdentificationModal(false);
+            }}
+            style={{
+              width: wp(7),
+              height: wp(7),
+              borderRadius: wp(7 / 2),
+              alignSelf: 'flex-end',
+              backgroundColor: Colors.lightBlue,
+              alignItems: 'center',
+              justifyContent: 'center',
+              marginTop: wp(3),
+              marginRight: wp(3),
+            }}
+          >
+            <FontAwesome name="close" color={Colors.white} size={19} />
+          </TouchableOpacity>
+          <View>
+            <View
+              style={{
+                marginLeft: wp(7),
+              }}
+            >
+              <Text style={{ ...styles.modalTitleText, fontSize: 18, fontFamily: Fonts.FiraSansRegular, color: 'grey', }}>Add Second Factor</Text>
+              <Text
+                style={{
+                  ...styles.modalInfoText,
+                  color: '#006DB4',
+                  fontFamily: Fonts.FiraSansRegular,
+                  fontSize: 18,
+                }}
+              >
+               {'F&F Identication'}
+              </Text>
+            </View>
+
+            <Text
+              style={{
+                ...styles.modalInfoText,
+                paddingTop: 8,
+                marginLeft: 30,
+                fontFamily: Fonts.FiraSansRegular,
+                fontSize: 14,
+              }}
+            >
+              {'Options depending on what details you have available for the contact'}
+            </Text>
+
+            <FlatList data={FNFIDENTIFICATIONDATA} renderItem={renderIdentificationItem} keyExtractor={(item) => item.id} />
+            <Text style={{
+              ...styles.modalInfoText,
+              paddingTop: 8,
+              marginLeft: 30,
+              fontFamily: Fonts.FiraSansRegular,
+              fontSize: 14,
+              marginTop: 50
+            }}>The recipient will be shown some characters of their phone number/ email that they need to confirm</Text>
+          </View>
+
+          <View style={{flexDirection:'row', alignItems: 'center'}}>
+          <TouchableOpacity
+              style={{...styles.btnContainer, marginTop: 30}}
+              onPress={() => {
+                console.log('Proceed clicked');
+              }}
+            >
+              <Text style={styles.btnText}>Proceed</Text>
+            </TouchableOpacity>
+
+            <TouchableOpacity
+              style={{ marginTop: 30, marginLeft: 40}}
+              onPress={() => {
+                console.log('Back clicked');
+                // navigation.goback();
+              }}
+            >
+              <Text style={{...styles.btnText, color:'#006DB4'}}>Back</Text>
+            </TouchableOpacity>
+          </View>
+        </View>
+      );
+    }
+
+
+    const SecretPhaseModal = () => {
+      return (
+        <View style={styles.modalContentContainer}>
+          <TouchableOpacity
+            activeOpacity={1}
+            onPress={() => {
+              setCustomSecretIdentificationModal(false);
+            }}
+            style={{
+              width: wp(7),
+              height: wp(7),
+              borderRadius: wp(7 / 2),
+              alignSelf: 'flex-end',
+              backgroundColor: Colors.lightBlue,
+              alignItems: 'center',
+              justifyContent: 'center',
+              marginTop: wp(3),
+              marginRight: wp(3),
+            }}
+          >
+            <FontAwesome name="close" color={Colors.white} size={19} />
+          </TouchableOpacity>
+          <View>
+            <View
+              style={{
+                marginLeft: wp(7),
+              }}
+            >
+              <Text style={{ ...styles.modalTitleText, fontSize: 18, fontFamily: Fonts.FiraSansRegular, color: 'grey' }}>Add Second Factor</Text>
+              <Text
+                style={{
+                  ...styles.modalInfoText,
+                  color: '#006DB4',
+                  fontFamily: Fonts.FiraSansRegular,
+                  fontSize: 18,
+                }}
+              >
+                {'Custom Secret Phrase'}
+              </Text>
+            </View>
+
+            <View style={styles.textInputContainer}>
+              <TextInput style={styles.textInput}  secureTextEntry={secretPhrase}/>
+              <TouchableWithoutFeedback
+                onPress={() => {
+                  setSecretPhrase(!secretPhrase);
+                }}
+              >
+                <Feather
+                  style={{
+                    marginLeft: 'auto',
+                  }}
+                  size={15}
+                  color={Colors.blue}
+                  name={secretPhrase ? 'eye-off' : 'eye'}
+                />
+              </TouchableWithoutFeedback>
+            </View>
+            <View style={styles.textInputContainer}>
+              <TextInput style={styles.textInput}  secureTextEntry={confirmSecretPhrase}/>
+              <TouchableWithoutFeedback
+                onPress={() => {
+                  setconfirmSecretPhrase(!confirmSecretPhrase);
+                }}
+              >
+                <Feather
+                  style={{
+                    marginLeft: 'auto',
+                  }}
+                  size={15}
+                  color={Colors.blue}
+                  name={confirmSecretPhrase ? 'eye-off' : 'eye'}
+                />
+              </TouchableWithoutFeedback>
+            </View>
+            <View style={styles.textInputContainer}>
+              <TextInput style={styles.textInput} placeholder="Add a hint" />
+            </View>
+            <Text
+              style={{
+                margin: 10,
+                marginLeft: 35,
+                color: '#6C6C6C',
+                width:'85%'
+              }}
+            >
+              The recipient will be shown the hint and they will have to provide the passphrase to accept the gift
+            </Text>
+
+            <TouchableOpacity
+              style={styles.btnContainer}
+              onPress={() => {
+                console.log('321');
+              }}
+            >
+              <Text style={styles.btnText}>Proceed</Text>
+            </TouchableOpacity>
+          </View>
+        </View>
+      );
+    }
+
 
 
   const renderButton = ( text, condn ) => {
@@ -130,6 +538,7 @@ const GiftDetails = ( { navigation } ) => {
     >
       <SafeAreaView style={styles.viewContainer}>
         <StatusBar backgroundColor={Colors.backgroundColor} barStyle="dark-content" />
+        <View style={styles.advancedButton}>
         <View
           style={[
             CommonStyles.headerContainer,
@@ -145,15 +554,58 @@ const GiftDetails = ( { navigation } ) => {
             }}
           >
             <View style={styles.headerLeftIconInnerContainer}>
-              {/* <FontAwesome
-                name="long-arrow-left"
-                color={Colors.blue}
-                size={17}
-              /> */}
               <LeftArrow />
             </View>
           </TouchableOpacity>
         </View>
+
+        <View>
+            <TouchableOpacity
+              onPress={() =>
+                setAdvanceSettingsModal( true )
+              }
+              style={{
+                height: 30,
+                width: 100,
+                justifyContent: 'center',
+                alignItems: 'center',
+                borderRadius: 8,
+                backgroundColor: Colors.blue,
+              }}
+            >
+              <Text style={{
+                color: 'white'
+              }}>{'Advanced'}</Text>
+            </TouchableOpacity>
+          </View>
+        </View>
+        {advanceSettingsModal && <ModalContainer
+          closeBottomSheet={() => setAdvanceSettingsModal( false )}
+          visible={advanceSettingsModal}
+          onBackground={()=>setAdvanceSettingsModal( false )}
+        >
+          {AdvancedSettingsModal()}
+        </ModalContainer>}
+
+        {FnFIdentificationModal && <ModalContainer
+          closeBottomSheet={() => setFnFIdentificationModal( false )}
+          visible={FnFIdentificationModal}
+          onBackground={()=>setFnFIdentificationModal( false )}
+        >
+          {FandFIndentificationModal()}
+        </ModalContainer>}
+
+        {
+          customSecretIdentificationModal && <ModalContainer
+            closeBottomSheet={() => {
+              setCustomSecretIdentificationModal( false )
+            }}
+            visible={customSecretIdentificationModal}
+            onBackground={()=>{setCustomSecretIdentificationModal( false )}}
+          >
+            {SecretPhaseModal()}
+          </ModalContainer>}
+
         <View
           style={{
             flexDirection: 'row',
@@ -440,19 +892,6 @@ const styles = StyleSheet.create( {
     fontFamily: Fonts.FiraSansRegular,
     marginHorizontal: wp( 2 )
   },
-  imageView: {
-    width: 18,
-    height: 18,
-    borderRadius: 9,
-    backgroundColor: Colors.white,
-    justifyContent: 'center',
-    alignItems: 'center',
-    elevation: 6,
-    shadowOpacity: 0.1,
-    shadowOffset: {
-      width: 1, height: 1
-    },
-  },
   dropdownBox: {
     flexDirection: 'row',
     borderColor: Colors.white,
@@ -727,19 +1166,19 @@ const styles = StyleSheet.create( {
     marginRight: 20,
     backgroundColor: Colors.white,
   },
-  inputBoxFocused: {
-    borderWidth: 0.5,
-    borderRadius: 10,
-    marginLeft: 20,
-    marginRight: 20,
-    elevation: 10,
-    shadowColor: Colors.borderColor,
-    shadowOpacity: 10,
-    shadowOffset: {
-      width: 10, height: 10
-    },
-    backgroundColor: Colors.white,
-  },
+  // inputBoxFocused: {
+  //   borderWidth: 0.5,
+  //   borderRadius: 10,
+  //   marginLeft: 20,
+  //   marginRight: 20,
+  //   elevation: 10,
+  //   shadowColor: Colors.borderColor,
+  //   shadowOpacity: 10,
+  //   shadowOffset: {
+  //     width: 10, height: 10
+  //   },
+  //   backgroundColor: Colors.white,
+  // },
   accImage:{
     marginRight: wp( 4 )
   },
@@ -775,6 +1214,71 @@ const styles = StyleSheet.create( {
   },
   headerLeftIconInnerContainer:{
     marginLeft: wp( 8.7 ),
+  },
+  cardContainer:{
+    flexDirection:'row',
+    alignItems:'center',
+    backgroundColor:'#fff',
+    width:'80%',
+    alignSelf:'center',
+    padding:13,
+    borderRadius:8,
+    paddingHorizontal:20,
+    marginVertical:10,
+  },
+  identificationHeading:{
+    color:'#006DB4',
+    fontSize:13,
+    fontWeight:'400',
+    fontFamily: Fonts.FiraSansRegular,
+
+  },
+  identificationDescription:{
+    color:'#6C6C6C',
+    fontSize:11,
+    fontWeight:'400',
+    width:225,
+    fontFamily: Fonts.FiraSansRegular,
+
+  },
+  radioBtnContainer:{
+    marginRight:10
+  },
+  advancedButton: {
+    flexDirection:'row',
+    alignItems:'center',
+    justifyContent:'space-between',
+    width: '94%'
+  },
+  textInputContainer:{
+    flexDirection:'row',
+    alignItems:'center',
+    justifyContent:'space-between',
+    width:'85%',
+    alignSelf:'center',
+    marginTop:10,
+    backgroundColor:'#fff',
+    padding:14,
+    borderRadius:10
+  },
+  textInput:{
+    width:'92%'
+  },
+  btnContainer:{
+    marginTop:10,
+    backgroundColor:'#006DB4',
+    width:100,
+    padding:14,
+    borderRadius:6,
+    marginLeft: 30,
+    justifyContent:'center',
+    alignItems:'center',
+  },
+  btnText: {
+    color: '#fff',
+    fontWeight:'500',
+    fontSize:15,
+    fontFamily: Fonts.FiraSansRegular
   }
 } )
